@@ -14,9 +14,10 @@
 		"preview": "next build && next start"
 	},
 	"dependencies": {
-		"@algolia/autocomplete-core": "1.19.0",
+		"@algolia/autocomplete-core": "1.19.1",
+		"babel-plugin-react-compiler": "^19.1.0-rc.1",
 		"clsx": "2.1.1",
-		"next": "^15.3.1",
+		"next": "^15.4.0-canary.20",
 		"react": "^19.1.0",
 		"react-dom": "^19.1.0",
 		"react-hotkeys-hook": "5.0.1",
@@ -26,9 +27,9 @@
 	"devDependencies": {
 		"@biomejs/biome": "1.9.4",
 		"@locator/babel-jsx": "^0.4.4",
-		"@types/node": "22.15.2",
+		"@types/node": "22.15.3",
 		"@types/react": "19.1.2",
-		"@types/react-dom": "19.1.2",
+		"@types/react-dom": "19.1.3",
 		"next-rspack": "^15.3.1",
 		"prettier": "^3.5.3",
 		"prettier-plugin-css-order": "^2.1.2",
